cleaned readme

+++ Makefile
@@ -1,6 +1,7 @@
-setup:
-	cp .local.env.copy .local.env
-
 install_package:
 	@pip uninstall -y rakugo || :
 	@pip install -e .
+
+setup_local:
+	test -f .local.env || cp .local.env.copy .local.env
+	mkdir -p local_data
